pyproject.toml
[project]
name = "hipsdb"
version = "0.1.0"
description = "A data server for HiPS data analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "blessings>=1.7",
    "click>=8.2.1",
    "django>=5.2.4",
    "django-click>=2.4.1",
    "django-extensions>=4.1",
]

[dependency-groups]
dev = [
    "hypothesis>=6.135.26",
    "ipython>=8.37.0",
    "pytest>=8.4.1",
    "ruff>=0.12.2",
]
